requirements.txt
django
django-crispy-forms
pillow
djangorestframework